+++ src/rules/layout-formatting.ts
@@ -1,4 +1,4 @@
 // https://eslint.org/docs/latest/rules/#layout--formatting
 export = {
   "line-comment-position": "error"
-} satisfies ESLintConfig["rules"];
+} satisfies ESLintRules;