+++ src/types.d.ts
@@ -2,4 +2,5 @@ import type { TSESLint } from "@typescript-eslint/utils";
 
 declare global {
   type ESLintConfig = TSESLint.ClassicConfig.Config;
+  type ESLintRules = TSESLint.ClassicConfig.RulesRecord;
 }